requirements.txt
numpy>=1.23
scipy>=1.10
sentence-transformers>=2.6
faiss-cpu>=1.7
openai>=1.0


 